refactor(ucmc-web): extract features/waivers/ from features/auth/
Waiver attestation is a compliance concern (paper-waiver lifecycle, cycle
rollover, version invalidation), not part of the auth identity surface.
Move the ~1,200 lines of waiver code out of features/auth/ and into a new
features/waivers/ so each feature has one user-facing job.

The requireCurrentWaiver guard stays in features/auth/guards.ts because
it composes auth + waiver state; the ESLint allowlist exposes
features/waivers/api/queries.ts and the type-only server/waiver-fns.ts
shell to features/auth only — every other feature is fully banned.
