+++ apps/web/eslint.config.js
@@ -101,6 +101,11 @@ export default [
             //    features/auth's public API (use-auth, view-mode,
             //    guards) is exempted because it's the foundational
             //    "who is the user" surface every feature needs.
+            //    features/waivers' public API (api/queries.ts +
+            //    server/waiver-fns.ts shell for type imports) is
+            //    exempted only for features/auth, because the
+            //    `requireCurrentWaiver` route guard composes auth +
+            //    waiver state and lives in features/auth/guards.ts.
             {
               target: "./src/features/announcements",
               from: "./src/features/auth",
@@ -115,12 +120,21 @@ export default [
               from: "./src/features/members",
             },
             {
+              target: "./src/features/announcements",
+              from: "./src/features/waivers",
+            },
+            {
               target: "./src/features/auth",
               from: "./src/features/announcements",
             },
             {
               target: "./src/features/auth",
               from: "./src/features/members",
+            },
+            {
+              target: "./src/features/auth",
+              from: "./src/features/waivers",
+              except: ["./api/queries.ts", "./server/waiver-fns.ts"],
             },
             {
               target: "./src/features/members",
@@ -134,6 +148,10 @@ export default [
             {
               target: "./src/features/members",
               from: "./src/features/announcements",
+            },
+            {
+              target: "./src/features/members",
+              from: "./src/features/waivers",
             },
             {
               target: "./src/features/announcements",
@@ -185,6 +203,10 @@ export default [
               from: "./src/features/audit",
             },
             {
+              target: "./src/features/waivers",
+              from: "./src/features/audit",
+            },
+            {
               target: "./src/features/audit",
               from: "./src/features/announcements",
             },
@@ -198,6 +220,30 @@ export default [
             },
             {
               target: "./src/features/audit",
+              from: "./src/features/members",
+            },
+            {
+              target: "./src/features/audit",
+              from: "./src/features/waivers",
+            },
+            {
+              target: "./src/features/feedback",
+              from: "./src/features/waivers",
+            },
+            {
+              target: "./src/features/waivers",
+              from: "./src/features/announcements",
+            },
+            {
+              target: "./src/features/waivers",
+              from: "./src/features/auth",
+            },
+            {
+              target: "./src/features/waivers",
+              from: "./src/features/feedback",
+            },
+            {
+              target: "./src/features/waivers",
               from: "./src/features/members",
             },
             // 2. Shared can't import features
