+++ apps/web/src/features/auth/api/query-keys.ts
@@ -11,24 +11,3 @@ export const PROFILE_QUERY_KEY = ["account", "profile"] as const;
 export const PASSKEY_LIST_QUERY_KEY = ["account", "passkeys"] as const;
 
 export const MY_EMAILS_QUERY_KEY = ["account", "emails"] as const;
-
-// Waiver attestation — the member's own status + history land under
-// "account", the officer queue under "members" so the cache trees mirror
-// the route trees that consume them.
-export const MY_WAIVER_STATUS_QUERY_KEY = [
-  "account",
-  "waiver",
-  "status",
-] as const;
-export const MY_WAIVER_HISTORY_QUERY_KEY = [
-  "account",
-  "waiver",
-  "history",
-] as const;
-export const WAIVER_PENDING_QUEUE_QUERY_KEY = [
-  "members",
-  "waivers",
-  "pending",
-] as const;
-export const waiverHistoryForUserQueryKey = (userId: string) =>
-  ["members", "waivers", "history", userId] as const;
